+++ gdsfactory/tests/test_paths.py
@@ -1,3 +1,4 @@
+import jsondiff
 import numpy as np
 import pytest
 from pytest_regressions.data_regression import DataRegressionFixture
@@ -25,9 +26,7 @@ def test_path():
     P.length()
     assert np.isclose(P.length(), 107.69901058617913), P.length()
 
-    X = gf.CrossSection()
-    X.add(width=1, offset=0, layer=0)
-
+    X = gf.CrossSection(width=1, offset=0, layer=(0, 0))
     c = gf.path.extrude(P, X)
     return c
 
@@ -36,13 +35,11 @@ def test_path():
 def rename():
     p = gf.path.arc()
 
-    # Create a blank CrossSection
-    X = gf.CrossSection()
-
-    # Add a a few "sections" to the cross-section
-    X.add(width=1, offset=0, layer=0, ports=("in", "out"))
-    X.add(width=3, offset=2, layer=2)
-    X.add(width=3, offset=-2, layer=2)
+    s1 = gf.Section(width=3, offset=2, layer=(2, 0))
+    s2 = gf.Section(width=3, offset=-2, layer=(2, 0))
+    X = gf.CrossSection(
+        width=1, offset=0, layer=(0, 0), port_names=("in", "out"), sections=[s1, s2]
+    )
 
     # Combine the Path and the CrossSection
     straight = gf.path.extrude(p, cross_section=X)
@@ -70,11 +67,16 @@ def double_loop():
     P.rotate(-45)
 
     # Create the crosssection
-    X = gf.CrossSection()
-    X.add(width=0.5, offset=2, layer=0)
-    X.add(width=0.5, offset=4, layer=1)
-    X.add(width=1.5, offset=0, layer=2, ports=["in", "out"])
-    X.add(width=1, offset=0, layer=3)
+    s1 = gf.Section(width=0.5, offset=2, layer=(0, 0))
+    s2 = gf.Section(width=0.5, offset=4, layer=(1, 0))
+    s3 = gf.Section(width=1, offset=0, layer=(3, 0))
+    X = gf.CrossSection(
+        width=1.5,
+        offset=0,
+        layer=(2, 0),
+        port_names=["in", "out"],
+        sections=[s1, s2, s3],
+    )
 
     c = gf.path.extrude(P, X, simplify=0.3)
     return c
@@ -83,16 +85,28 @@ def double_loop():
 @cell
 def transition():
     c = gf.Component()
-    X1 = gf.CrossSection()
-    X1.add(width=1.2, offset=0, layer=2, name="wg", ports=("in1", "out1"))
-    X1.add(width=2.2, offset=0, layer=3, name="etch")
-    X1.add(width=1.1, offset=3, layer=1, name="wg2")
+    s1 = gf.Section(width=2.2, offset=0, layer=(3, 0), name="etch")
+    s2 = gf.Section(width=1.1, offset=3, layer=(1, 0), name="wg2")
+    X1 = gf.CrossSection(
+        width=1.2,
+        offset=0,
+        layer=(2, 0),
+        name="wg",
+        port_names=("in1", "out1"),
+        sections=[s1, s2],
+    )
 
     # Create the second CrossSection that we want to transition to
-    X2 = gf.CrossSection()
-    X2.add(width=1, offset=0, layer=2, name="wg", ports=("in1", "out1"))
-    X2.add(width=3.5, offset=0, layer=3, name="etch")
-    X2.add(width=3, offset=5, layer=1, name="wg2")
+    s1 = gf.Section(width=3.5, offset=0, layer=(3, 0), name="etch")
+    s2 = gf.Section(width=3, offset=5, layer=(1, 0), name="wg2")
+    X2 = gf.CrossSection(
+        width=1,
+        offset=0,
+        layer=(2, 0),
+        name="wg",
+        port_names=("in1", "out1"),
+        sections=[s1, s2],
+    )
 
     Xtrans = gf.path.transition(cross_section1=X1, cross_section2=X2, width_type="sine")
 
@@ -146,8 +160,9 @@ def test_settings(component: Component, data_regression: DataRegressionFixture) 
 
 def test_layers1():
     P = gf.path.straight(length=10.001)
-    X = gf.CrossSection()
-    X.add(width=0.5, offset=0, layer=gf.LAYER.SLAB90, ports=["in", "out"])
+    X = gf.CrossSection(
+        width=0.5, offset=0, layer=gf.LAYER.SLAB90, port_names=["in", "out"]
+    )
     c = gf.path.extrude(P, X, simplify=5e-3)
     assert c.ports["in"].layer == gf.LAYER.SLAB90
     assert c.ports["out"].position[0] == 10.001, c.ports["out"].position[0]
@@ -164,21 +179,73 @@ def test_layers2():
 
 
 def test_copy():
-    X = gf.CrossSection()
-    X.add(width=0.5, offset=0, layer=gf.LAYER.SLAB90, ports=["in", "out"])
-    x2 = X.copy()
-    assert x2
+    x1 = gf.CrossSection(
+        width=0.5, offset=0, layer=gf.LAYER.SLAB90, port_names=["in", "out"]
+    )
+    x2 = x1.copy()
+
+    d = jsondiff.diff(x1.dict(), x2.dict())
+    assert len(d) == 0, d
 
 
 if __name__ == "__main__":
-    c = transition()
+    # c = transition()
     # c = test_path()
     # print(c.name)
 
-    # test_copy()
+    test_copy()
     # c = test_layers2()
     # c = transition()
     # c = double_loop()
     # c = rename()
     # c.pprint()
-    c.show()
+
+    # c.show()
+
+    # c = gf.Component()
+    # s1 = gf.Section(width=2.2, offset=0, layer=(3, 0), name="etch")
+    # s2 = gf.Section(width=1.1, offset=3, layer=(1, 0), name="wg2")
+    # X1 = gf.CrossSection(
+    #     width=1.2,
+    #     offset=0,
+    #     layer=(2, 0),
+    #     name="wg",
+    #     port_names=("in1", "out1"),
+    #     sections=[s1, s2],
+    # )
+
+    # # Create the second CrossSection that we want to transition to
+    # s1 = gf.Section(width=3.5, offset=0, layer=(3, 0), name="etch")
+    # s2 = gf.Section(width=3, offset=5, layer=(1, 0), name="wg2")
+    # X2 = gf.CrossSection(
+    #     width=1,
+    #     offset=0,
+    #     layer=(2, 0),
+    #     name="wg",
+    #     port_names=("in1", "out1"),
+    #     sections=[s1, s2],
+    # )
+
+    # Xtrans = gf.path.transition(cross_section1=X1, cross_section2=X2, width_type="sine")
+
+    # P1 = gf.path.straight(length=5)
+    # P2 = gf.path.straight(length=5)
+
+    # wg1 = gf.path.extrude(P1, X1)
+    # wg2 = gf.path.extrude(P2, X2)
+
+    # P4 = gf.path.euler(radius=25, angle=45, p=0.5, use_eff=False)
+    # wg_trans = gf.path.extrude(P4, Xtrans)
+
+    # # print("wg1", wg1)
+    # # print("wg2", wg2)
+    # # print("wg3", wg_trans)
+    # # wg_trans.pprint()
+
+    # wg1_ref = c << wg1
+    # wgt_ref = c << wg_trans
+    # wgt_ref.connect("in1", wg1_ref.ports["out1"])
+
+    # wg2_ref = c << wg2
+    # wg2_ref.connect("in1", wgt_ref.ports["out1"])
+    # c.show()