+++ gdsfactory/tests/test_paths_extrude.py
@@ -4,14 +4,15 @@ from gdsfactory.tech import LAYER
 
 @gf.cell
 def test_path_extrude_multiple_ports() -> gf.Component:
-    X = gf.CrossSection()
-    X.add(width=0.5, offset=0, layer=LAYER.SLAB90, ports=["o1", "o2"])
-    X.add(
+    s = gf.Section(
         width=2.0,
         offset=-4,
         layer=LAYER.HEATER,
-        ports=["e1", "e2"],
+        port_names=["e1", "e2"],
         port_types=("electrical", "electrical"),
+    )
+    X = gf.CrossSection(
+        width=0.5, offset=0, layer=LAYER.SLAB90, port_names=["o1", "o2"], sections=[s]
     )
     P = gf.path.straight(npoints=100, length=10)
     c = gf.path.extrude(P, X)